# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nokta Platform API
  version: 1.0.0
  description: >-
    Consolidated API surface for Nokta POS, call center, billing, feature flag, and onboarding services.
servers:
  - url: https://api.nokta-pos.com/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
security:
  - bearerAuth: []
  - TenantHeader: []
paths:
...
  /call-center/queue:
    get:
      summary: Fetch the current call queue snapshot
      tags: [Call Center]
      parameters:
        - in: query
          name: tenantId
          schema:
            type: integer
          required: false
          description: Override tenant context when requesting cross-tenant insights
      responses:
        '200':
          description: Active and waiting calls
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenantId:
                    type: integer
                  count:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/CallQueueEntry'
        '403':
          description: Feature flag disabled
  /call-center/dashboard:
    get:
      summary: Retrieve live KPIs for the call center
      tags: [Call Center]
      parameters:
        - $ref: '#/components/parameters/Range'
        - $ref: '#/components/parameters/BranchId'
      responses:
        '200':
          description: KPI payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallCenterMetrics'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
  parameters:
    Range:
      in: query
      name: range
      schema:
        type: string
        enum: [today, 7d, 30d]
    BranchId:
      in: query
      name: branchId
      schema:
        type: integer
...
  schemas:
...
    CallQueueEntry:
      type: object
      properties:
        id:
          type: string
        callerNumber:
          type: string
        displayName:
          type: string
        priority:
          type: integer
        status:
          type: string
        waitingSince:
          type: string
          format: date-time
        customerId:
          type: integer
          nullable: true
        lastOrderId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        preferredBranchId:
          type: integer
          nullable: true
        loyaltyPoints:
          type: integer
        lastOrderAt:
          type: string
          format: date-time
          nullable: true
    CallCenterMetrics:
      type: object
      properties:
        queueLength:
          type: integer
        activeCalls:
          type: integer
        averageWaitTimeSeconds:
          type: integer
        slaBreaches:
          type: integer
        ordersCreated:
          type: integer
        serviceLevelAchievement:
          type: integer
        updatedAt:
          type: string
          format: date-time
...